# openBanking/src/main/webapp/META-INF/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: ServiceDesk@openbanking.org.uk
    name: Service Desk
  description: Swagger for Payment Initiation API Specification
  license:
    name: open-licence
    url: https://www.openbanking.org.uk/open-licence
  termsOfService: https://www.openbanking.org.uk/terms
  title: Payment Initiation API
  version: 3.1.10
servers:
- url: /
- url: /open-banking/v3.1/pisp
paths:
...
  /domestic-payments:
    post:
      operationId: CreateDomesticPayments
      parameters:
      - description: "The time when the PSU last logged in with the TPP. \nAll dates\
          \ in the HTTP headers are represented as RFC 7231 Full Dates. An example\
          \ is below: \nSun, 10 Sep 2017 19:43:31 UTC"
        explode: false
        in: header
        name: x-fapi-auth-date
        required: false
        schema:
          pattern: "^(Mon|Tue|Wed|Thu|Fri|Sat|Sun), \\d{2} (Jan|Feb|Mar|Apr|May|Jun|Jul|Aug|Sep|Oct|Nov|Dec)\
            \ \\d{4} \\d{2}:\\d{2}:\\d{2} (GMT|UTC)$"
          type: string
        style: simple
      - description: The PSU's IP address if the PSU is currently logged in with the
          TPP.
        explode: false
        in: header
        name: x-fapi-customer-ip-address
        required: false
        schema:
          type: string
        style: simple
      - description: An RFC4122 UID used as a correlation id.
        explode: false
        in: header
        name: x-fapi-interaction-id
        required: false
        schema:
          type: string
        style: simple
      - description: An Authorisation Token as per https://tools.ietf.org/html/rfc6750
        explode: false
        in: header
        name: Authorization
        required: true
        schema:
          type: string
        style: simple
      - description: |
          Every request will be processed only once per x-idempotency-key.  The
          Idempotency Key will be valid for 24 hours.
        explode: false
        in: header
        name: x-idempotency-key
        required: true
        schema:
          maxLength: 40
          pattern: ^(?!\s)(.*)(\S)$
          type: string
        style: simple
      - description: A detached JWS signature of the body of the payload.
        explode: false
        in: header
        name: x-jws-signature
        required: true
        schema:
          type: string
        style: simple
      - description: Indicates the user-agent that the PSU is using.
        explode: false
        in: header
        name: x-customer-user-agent
        required: false
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json; charset=utf-8:
            schema:
              $ref: '#/components/schemas/OBWriteDomestic2'
          application/json:
            schema:
              $ref: '#/components/schemas/OBWriteDomestic2'
          application/jose+jwe:
            schema:
              $ref: '#/components/schemas/OBWriteDomestic2'
        description: Default
        required: true
      responses:
        "201":
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/OBWriteDomesticResponse5'
            application/json:
              schema:
                $ref: '#/components/schemas/OBWriteDomesticResponse5'
            application/jose+jwe:
              schema:
                $ref: '#/components/schemas/OBWriteDomesticResponse5'
          description: Domestic Payments Created
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
            x-jws-signature:
              description: |
                Header containing a detached JWS signature of the body of the payload.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "400":
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
            application/json:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
            application/jose+jwe:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
          description: Bad request
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
            x-jws-signature:
              description: |
                Header containing a detached JWS signature of the body of the payload.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "401":
          description: Unauthorized
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "403":
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
            application/json:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
            application/jose+jwe:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
          description: Forbidden
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
            x-jws-signature:
              description: |
                Header containing a detached JWS signature of the body of the payload.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "405":
          description: Method Not Allowed
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "406":
          description: Not Acceptable
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "415":
          description: Unsupported Media Type
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "429":
          description: Too Many Requests
          headers:
            Retry-After:
              description: Number in seconds to wait
              explode: false
              schema:
                type: integer
              style: simple
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "500":
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
            application/json:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
            application/jose+jwe:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
          description: Internal Server Error
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
            x-jws-signature:
              description: |
                Header containing a detached JWS signature of the body of the payload.
              explode: false
              required: true
              schema:
                type: string
              style: simple
      security:
      - PSUOAuth2Security:
        - payments
      summary: Create Domestic Payments
      tags:
      - Domestic Payments
      x-contentType: application/json; charset=utf-8
      x-accepts: application/json
      x-tags:
      - tag: Domestic Payments
  /domestic-payments/{DomesticPaymentId}:
    get:
      operationId: GetDomesticPaymentsDomesticPaymentId
      parameters:
      - description: DomesticPaymentId
        explode: false
        in: path
        name: DomesticPaymentId
        required: true
        schema:
          type: string
        style: simple
      - description: "The time when the PSU last logged in with the TPP. \nAll dates\
          \ in the HTTP headers are represented as RFC 7231 Full Dates. An example\
          \ is below: \nSun, 10 Sep 2017 19:43:31 UTC"
        explode: false
        in: header
        name: x-fapi-auth-date
        required: false
        schema:
          pattern: "^(Mon|Tue|Wed|Thu|Fri|Sat|Sun), \\d{2} (Jan|Feb|Mar|Apr|May|Jun|Jul|Aug|Sep|Oct|Nov|Dec)\
            \ \\d{4} \\d{2}:\\d{2}:\\d{2} (GMT|UTC)$"
          type: string
        style: simple
      - description: The PSU's IP address if the PSU is currently logged in with the
          TPP.
        explode: false
        in: header
        name: x-fapi-customer-ip-address
        required: false
        schema:
          type: string
        style: simple
      - description: An RFC4122 UID used as a correlation id.
        explode: false
        in: header
        name: x-fapi-interaction-id
        required: false
        schema:
          type: string
        style: simple
      - description: An Authorisation Token as per https://tools.ietf.org/html/rfc6750
        explode: false
        in: header
        name: Authorization
        required: true
        schema:
          type: string
        style: simple
      - description: Indicates the user-agent that the PSU is using.
        explode: false
        in: header
        name: x-customer-user-agent
        required: false
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/OBWriteDomesticResponse5'
            application/json:
              schema:
                $ref: '#/components/schemas/OBWriteDomesticResponse5'
            application/jose+jwe:
              schema:
                $ref: '#/components/schemas/OBWriteDomesticResponse5'
          description: Domestic Payments Read
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
            x-jws-signature:
              description: |
                Header containing a detached JWS signature of the body of the payload.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "400":
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
            application/json:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
            application/jose+jwe:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
          description: Bad request
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
            x-jws-signature:
              description: |
                Header containing a detached JWS signature of the body of the payload.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "401":
          description: Unauthorized
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "403":
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
            application/json:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
            application/jose+jwe:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
          description: Forbidden
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
            x-jws-signature:
              description: |
                Header containing a detached JWS signature of the body of the payload.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "405":
          description: Method Not Allowed
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "406":
          description: Not Acceptable
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "429":
          description: Too Many Requests
          headers:
            Retry-After:
              description: Number in seconds to wait
              explode: false
              schema:
                type: integer
              style: simple
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "500":
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
            application/json:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
            application/jose+jwe:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
          description: Internal Server Error
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
            x-jws-signature:
              description: |
                Header containing a detached JWS signature of the body of the payload.
              explode: false
              required: true
              schema:
                type: string
              style: simple
      security:
      - TPPOAuth2Security:
        - payments
      summary: Get Domestic Payments
      tags:
      - Domestic Payments
      x-accepts: application/json
      x-tags:
      - tag: Domestic Payments
components:
...
  schemas:
...
    Links:
      additionalProperties: false
      description: Links relevant to the payload
      example:
        Last: https://openapi-generator.tech
        Prev: https://openapi-generator.tech
        Next: https://openapi-generator.tech
        Self: https://openapi-generator.tech
        First: https://openapi-generator.tech
      properties:
        Self:
          format: uri
          type: string
        First:
          format: uri
          type: string
        Prev:
          format: uri
          type: string
        Next:
          format: uri
          type: string
        Last:
          format: uri
          type: string
      required:
      - Self
      type: object
    Meta:
      additionalProperties: false
      description: Meta Data relevant to the payload
      example:
        LastAvailableDateTime: 2000-01-23T04:56:07.000+00:00
        FirstAvailableDateTime: 2000-01-23T04:56:07.000+00:00
        TotalPages: 0
      properties:
        TotalPages:
          format: int32
          type: integer
        FirstAvailableDateTime:
          description: "All dates in the JSON payloads are represented in ISO 8601\
            \ date-time format. \nAll date-time fields in responses must include the\
            \ timezone. An example is below:\n2017-04-05T10:43:07+00:00"
          format: date-time
          type: string
        LastAvailableDateTime:
          description: "All dates in the JSON payloads are represented in ISO 8601\
            \ date-time format. \nAll date-time fields in responses must include the\
            \ timezone. An example is below:\n2017-04-05T10:43:07+00:00"
          format: date-time
          type: string
      title: MetaData
      type: object
...
    OBErrorResponse1:
      additionalProperties: false
      description: "An array of detail error codes, and messages, and URLs to documentation\
        \ to help remediation."
      properties:
        Code:
          description: "High level textual error code, to help categorize the errors."
          maxLength: 40
          minLength: 1
          type: string
        Id:
          description: "A unique reference for the error instance, for audit purposes,\
            \ in case of unknown/unclassified errors."
          maxLength: 40
          minLength: 1
          type: string
        Message:
          description: "Brief Error message, e.g., 'There is something wrong with\
            \ the request parameters provided'"
          maxLength: 500
          minLength: 1
          type: string
        Errors:
          items:
            $ref: '#/components/schemas/OBError1'
          minItems: 1
          type: array
      required:
      - Code
      - Errors
      - Message
      type: object
...
    OBExternalExtendedAccountType1Code:
      description: Specifies the extended type of account.
      enum:
      - Business
      - BusinessSavingsAccount
      - Charity
      - Collection
      - Corporate
      - Ewallet
      - Government
      - Investment
      - ISA
      - JointPersonal
      - Pension
      - Personal
      - PersonalSavingsAccount
      - Premier
      - Wealth
      type: string
...
    OBRisk1:
      additionalProperties: false
      description: |-
        The Risk section is sent by the initiating party to the ASPSP.
        It is used to specify additional details for risk scoring for Payments.
      example:
        PaymentPurposeCode: PaymentPurposeCode
        BeneficiaryPrepopulatedIndicator: true
        ContractPresentInidicator: true
        PaymentContextCode: BillingGoodsAndServicesInAdvance
        DeliveryAddress:
          StreetName: StreetName
          CountrySubDivision: CountrySubDivision
          AddressLine:
          - AddressLine
          - AddressLine
          BuildingNumber: BuildingNumber
          TownName: TownName
          Country: Country
          PostCode: PostCode
        MerchantCategoryCode: MerchantCategoryCode
        MerchantCustomerIdentification: MerchantCustomerIdentification
        BeneficiaryAccountType: null
      properties:
        PaymentContextCode:
          description: |
            Specifies the payment context
            * BillPayment - @deprecated
            * EcommerceGoods - @deprecated
            * EcommerceServices - @deprecated
            * Other - @deprecated
            * PartyToParty - @deprecated
          enum:
          - BillingGoodsAndServicesInAdvance
          - BillingGoodsAndServicesInArrears
          - PispPayee
          - EcommerceMerchantInitiatedPayment
          - FaceToFacePointOfSale
          - TransferToSelf
          - TransferToThirdParty
          - BillPayment
          - EcommerceGoods
          - EcommerceServices
          - Other
          - PartyToParty
          type: string
        MerchantCategoryCode:
          description: "Category code conform to ISO 18245, related to the type of\
            \ services or goods the merchant provides for the transaction."
          maxLength: 4
          minLength: 3
          type: string
        MerchantCustomerIdentification:
          description: The unique customer identifier of the PSU with the merchant.
          maxLength: 70
          minLength: 1
          type: string
        ContractPresentInidicator:
          description: Indicates if Payee has a contractual relationship with the
            PISP.
          type: boolean
        BeneficiaryPrepopulatedIndicator:
          description: Indicates if PISP has immutably prepopulated payment details
            in for the PSU.
          type: boolean
        PaymentPurposeCode:
          description: "Category code, related to the type of services or goods that\
            \ corresponds to the underlying purpose of the payment that conforms to\
            \ Recommended UK Purpose Code in ISO 20022 Payment Messaging List"
          maxLength: 4
          minLength: 3
          type: string
        BeneficiaryAccountType:
          $ref: '#/components/schemas/OBExternalExtendedAccountType1Code'
        DeliveryAddress:
          $ref: '#/components/schemas/OBRisk1_DeliveryAddress'
      type: object
...
    OBWriteDomestic2:
      additionalProperties: false
      example:
        Risk:
          PaymentPurposeCode: PaymentPurposeCode
          BeneficiaryPrepopulatedIndicator: true
          ContractPresentInidicator: true
          PaymentContextCode: BillingGoodsAndServicesInAdvance
          DeliveryAddress:
            StreetName: StreetName
            CountrySubDivision: CountrySubDivision
            AddressLine:
            - AddressLine
            - AddressLine
            BuildingNumber: BuildingNumber
            TownName: TownName
            Country: Country
            PostCode: PostCode
          MerchantCategoryCode: MerchantCategoryCode
          MerchantCustomerIdentification: MerchantCustomerIdentification
          BeneficiaryAccountType: null
        Data:
          ConsentId: ConsentId
          Initiation:
            SupplementaryData:
              key: "{}"
            LocalInstrument: null
            DebtorAccount:
              SecondaryIdentification: SecondaryIdentification
              SchemeName: null
              Identification: Identification
              Name: Name
            RemittanceInformation:
              Unstructured: Unstructured
              Reference: Reference
            EndToEndIdentification: EndToEndIdentification
            InstructionIdentification: InstructionIdentification
            CreditorAccount:
              SecondaryIdentification: SecondaryIdentification
              SchemeName: null
              Identification: Identification
              Name: Name
            InstructedAmount:
              Amount: Amount
              Currency: Currency
            CreditorPostalAddress:
              StreetName: StreetName
              CountrySubDivision: CountrySubDivision
              Department: Department
              AddressLine:
              - AddressLine
              - AddressLine
              - AddressLine
              - AddressLine
              - AddressLine
              BuildingNumber: BuildingNumber
              TownName: TownName
              Country: Country
              SubDepartment: SubDepartment
              AddressType: null
              PostCode: PostCode
      properties:
        Data:
          $ref: '#/components/schemas/OBWriteDomestic2_Data'
        Risk:
          $ref: '#/components/schemas/OBRisk1'
      required:
      - Data
      - Risk
      type: object
...
    OBWriteDomesticResponse5:
      additionalProperties: false
      example:
        Meta:
          LastAvailableDateTime: 2000-01-23T04:56:07.000+00:00
          FirstAvailableDateTime: 2000-01-23T04:56:07.000+00:00
          TotalPages: 0
        Links:
          Last: https://openapi-generator.tech
          Prev: https://openapi-generator.tech
          Next: https://openapi-generator.tech
          Self: https://openapi-generator.tech
          First: https://openapi-generator.tech
        Data:
          DomesticPaymentId: DomesticPaymentId
          Status: AcceptedCreditSettlementCompleted
          StatusUpdateDateTime: 2000-01-23T04:56:07.000+00:00
          Debtor:
            SecondaryIdentification: SecondaryIdentification
            SchemeName: SchemeName
            Identification: Identification
            Name: Name
          CreationDateTime: 2000-01-23T04:56:07.000+00:00
          ExpectedExecutionDateTime: 2000-01-23T04:56:07.000+00:00
          Refund:
            Account:
              SecondaryIdentification: SecondaryIdentification
              SchemeName: null
              Identification: Identification
              Name: Name
          Charges:
          - Type: null
            Amount:
              Amount: Amount
              Currency: Currency
            ChargeBearer: null
          - Type: null
            Amount:
              Amount: Amount
              Currency: Currency
            ChargeBearer: null
          ConsentId: ConsentId
          Initiation:
            SupplementaryData:
              key: "{}"
            LocalInstrument: null
            DebtorAccount:
              SecondaryIdentification: SecondaryIdentification
              SchemeName: null
              Identification: Identification
              Name: Name
            RemittanceInformation:
              Unstructured: Unstructured
              Reference: Reference
            EndToEndIdentification: EndToEndIdentification
            InstructionIdentification: InstructionIdentification
            CreditorAccount:
              SecondaryIdentification: SecondaryIdentification
              SchemeName: null
              Identification: Identification
              Name: Name
            InstructedAmount:
              Amount: Amount
              Currency: Currency
            CreditorPostalAddress:
              StreetName: StreetName
              CountrySubDivision: CountrySubDivision
              Department: Department
              AddressLine:
              - AddressLine
              - AddressLine
              - AddressLine
              - AddressLine
              - AddressLine
              BuildingNumber: BuildingNumber
              TownName: TownName
              Country: Country
              SubDepartment: SubDepartment
              AddressType: null
              PostCode: PostCode
          ExpectedSettlementDateTime: 2000-01-23T04:56:07.000+00:00
          MultiAuthorisation:
            Status: Authorised
            NumberReceived: 6
            LastUpdateDateTime: 2000-01-23T04:56:07.000+00:00
            ExpirationDateTime: 2000-01-23T04:56:07.000+00:00
            NumberRequired: 0
      properties:
        Data:
          $ref: '#/components/schemas/OBWriteDomesticResponse5_Data'
        Links:
          $ref: '#/components/schemas/Links'
        Meta:
          $ref: '#/components/schemas/Meta'
      required:
      - Data
      type: object
...
    OBWriteDomestic2_Data:
      example:
        ConsentId: ConsentId
        Initiation:
          SupplementaryData:
            key: "{}"
          LocalInstrument: null
          DebtorAccount:
            SecondaryIdentification: SecondaryIdentification
            SchemeName: null
            Identification: Identification
            Name: Name
          RemittanceInformation:
            Unstructured: Unstructured
            Reference: Reference
          EndToEndIdentification: EndToEndIdentification
          InstructionIdentification: InstructionIdentification
          CreditorAccount:
            SecondaryIdentification: SecondaryIdentification
            SchemeName: null
            Identification: Identification
            Name: Name
          InstructedAmount:
            Amount: Amount
            Currency: Currency
          CreditorPostalAddress:
            StreetName: StreetName
            CountrySubDivision: CountrySubDivision
            Department: Department
            AddressLine:
            - AddressLine
            - AddressLine
            - AddressLine
            - AddressLine
            - AddressLine
            BuildingNumber: BuildingNumber
            TownName: TownName
            Country: Country
            SubDepartment: SubDepartment
            AddressType: null
            PostCode: PostCode
      properties:
        ConsentId:
          description: "OB: Unique identification as assigned by the ASPSP to uniquely\
            \ identify the consent resource."
          maxLength: 128
          minLength: 1
          type: string
        Initiation:
          $ref: '#/components/schemas/OBWriteDomestic2_Data_Initiation'
      required:
      - ConsentId
      - Initiation
      type: object
...
  securitySchemes:
    TPPOAuth2Security:
      description: TPP client credential authorisation flow with the ASPSP
      flows:
        clientCredentials:
          scopes:
            payments: Generic payment scope
          tokenUrl: https://authserver.example/token
      type: oauth2
    PSUOAuth2Security:
      description: "OAuth flow, it is required when the PSU needs to perform SCA with\
        \ the ASPSP when a TPP wants to access an ASPSP resource owned by the PSU"
      flows:
        authorizationCode:
          authorizationUrl: https://authserver.example/authorization
          scopes:
            payments: Generic payment scope
          tokenUrl: https://authserver.example/token
      type: oauth2